requirements.txt
cvzone==1.5.6
Flask==2.2.3
Pillow==9.4.0
Pillow==9.5.0
pytesseract==0.3.10
Werkzeug==2.2.3
